webpack: Fix CSS Modules
I go back and forth between CSS modules making me more productive or
less productive :D. This commit splits the CSS imports so that our CSS
imports are modules but CSS imported from a node module can be imported
generically. This is to support components like SwaggerUI that require a
lot of CSS and don't make it super easy to use the module.