bugfix: List padding.
This commit fixes an issue where there was a lot of space between the
proceeding paragraph and the start of the referenced list. To do so, I
swapped out the paradigm of "margin bottom" to use a "margin top"
structure so that images, code blocks, etc would put the margin at the
top and expect the proceeding element to put the margin on top as well.
To fix headings, I substracted the margin so there wasn't a large gap
between a heading and the first bit of content.