github: Add build steps. (#76)
This commit adds build steps. Technically, Netlify is responsible for
this at the moment. But it's not clear if everyone will be able to get a
Netlify build from our community, so I want to add these to ensure we
can block on failed builds regardless of if Netlify will provide a link
or not.