feat: Add console to tidbyt.dev. (#112)
This commit adds a serial console to tidbyt.dev. This is very similar to
console.tidbyt.com with only slight additions. One is that we check for
vender and product ID, another is that we have a download button. But
the main benefit is we can extend this console to do a lot more, like be
able to flash a Tidbyt or login to publish the log results.